Add Authentication Protocol, Move to `dev.gmitch215` Namespace
