--- src/android/main/xyz/gmitch215/tabroom/util/http.android.kt
@@ -1,18 +1,0 @@
-package xyz.gmitch215.tabroom.util
-
-import io.ktor.client.engine.*
-import io.ktor.client.engine.okhttp.*
-import kotlinx.coroutines.Dispatchers
-import java.util.concurrent.TimeUnit
-
-internal actual val engine: HttpClientEngine = OkHttp.create {
-    pipelining = true
-    dispatcher = Dispatchers.IO.limitedParallelism(PARALLEL_COUNT)
-
-    config {
-        followRedirects(true)
-        connectTimeout(30, TimeUnit.SECONDS)
-        readTimeout(30, TimeUnit.SECONDS)
-        writeTimeout(30, TimeUnit.SECONDS)
-    }
-}
