--- src/jvm/main/xyz/gmitch215/tabroom/api/fetcher.kt
@@ -1,12 +1,0 @@
-@file:JvmName("TabroomAPI")
-
-package xyz.gmitch215.tabroom.api
-
-import kotlinx.coroutines.runBlocking
-
-/**
- * Gets a tournament by its ID.
- * @param id The ID of the tournament.
- * @return The tournament.
- */
-fun getTournamentSync(id: Int): Tournament = runBlocking { getTournament(id) }
