--- src/mingw/main/xyz/gmitch215/tabroom/util/http.mingw.kt
@@ -1,14 +1,0 @@
-package xyz.gmitch215.tabroom.util
-
-import io.ktor.client.engine.*
-import io.ktor.client.engine.winhttp.*
-import io.ktor.http.*
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.IO
-import xyz.gmitch215.tabroom.util.PARALLEL_COUNT
-
-internal actual val engine: HttpClientEngine = WinHttp.create {
-    pipelining = true
-    protocolVersion = HttpProtocolVersion.HTTP_2_0
-    dispatcher = Dispatchers.IO.limitedParallelism(PARALLEL_COUNT)
-}
