+++ src/native/main/dev/gmitch215/tabroom/util/http.native.kt
@@ -1,0 +1,27 @@
+@file:OptIn(ExperimentalForeignApi::class, UnsafeNumber::class)
+
+package dev.gmitch215.tabroom.util
+
+import kotlinx.cinterop.ByteVar
+import kotlinx.cinterop.ExperimentalForeignApi
+import kotlinx.cinterop.UnsafeNumber
+import kotlinx.cinterop.allocArray
+import kotlinx.cinterop.memScoped
+import platform.posix.snprintf
+
+actual fun encodeURL(url: String): String = url.encodeToByteArray().joinToString("") {
+    memScoped {
+        val str = allocArray<ByteVar>(4) // 3 for %XX + 1 for null terminator
+        snprintf(str, 4.toUInt(), "%%%02X", it)
+        str.toString()
+    }
+}
+
+actual fun decodeURL(url: String): String {
+    val decoded = url.replace("%[0-9A-Fa-f]{2}".toRegex()) { match ->
+        val hex = match.value.substring(1)
+        val byte = hex.toInt(16).toByte()
+        byte.toInt().toChar().toString()
+    }
+    return decoded
+}
